% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manufacturers_functions.R
\name{accounts.products.get}
\alias{accounts.products.get}
\title{Gets the product from a Manufacturer Center account, including productissues.}
\usage{
accounts.products.get(parent, name)
}
\arguments{
\item{parent}{Parent ID in the format `accounts/{account_id}`}

\item{name}{Name in the format `{target_country}:{content_language}:{product_id}`}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/manufacturercenter
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/manufacturercenter)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/manufacturers/}{Google Documentation}
}
