% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manufacturers_functions.R
\name{accounts.products.list}
\alias{accounts.products.list}
\title{Lists all the products in a Manufacturer Center account.}
\usage{
accounts.products.list(parent, pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{parent}{Parent ID in the format `accounts/{account_id}`}

\item{pageToken}{The token returned by the previous request}

\item{pageSize}{Maximum number of product statuses to return in the response, used for}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/manufacturercenter
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/manufacturercenter)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/manufacturers/}{Google Documentation}
}
