% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manufacturers_objects.R
\name{Attributes}
\alias{Attributes}
\title{Attributes Object}
\usage{
Attributes(productLine = NULL, brand = NULL, productName = NULL,
  productType = NULL, productPageUrl = NULL, mpn = NULL, gtin = NULL,
  title = NULL)
}
\arguments{
\item{productLine}{The name of the group of products related to the product}

\item{brand}{The brand name of the product}

\item{productName}{The canonical name of the product}

\item{productType}{The manufacturer's category of the product}

\item{productPageUrl}{The URL of the manufacturer's detail page of the product}

\item{mpn}{The Manufacturer Part Number (MPN) of the product}

\item{gtin}{The Global Trade Item Number (GTIN) of the product}

\item{title}{The title of the product}
}
\value{
Attributes object
}
\description{
Attributes Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Attributes of the product. For more information, seehttps://support.google.com/manufacturers/answer/6124116.
}
