% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manufacturers_objects.R
\name{Issue}
\alias{Issue}
\title{Issue Object}
\usage{
Issue(attribute = NULL, severity = NULL, description = NULL,
  type = NULL)
}
\arguments{
\item{attribute}{If present, the attribute that triggered the issue}

\item{severity}{The severity of the issue}

\item{description}{Description of the issue}

\item{type}{The server-generated type of the issue, for example,}
}
\value{
Issue object
}
\description{
Issue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Product issue.
}
