% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manufacturers_objects.R
\name{ListProductsResponse}
\alias{ListProductsResponse}
\title{Manufacturer Center API Objects 
Public API for managing Manufacturer Center related data.}
\usage{
ListProductsResponse(nextPageToken = NULL, products = NULL)
}
\arguments{
\item{nextPageToken}{The token for the retrieval of the next page of product statuses}

\item{products}{List of the products}
}
\value{
ListProductsResponse object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:01:47
filename: /Users/mark/dev/R/autoGoogleAPI/googlemanufacturersv1.auto/R/manufacturers_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ListProductsResponse Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
