% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manufacturers_functions.R
\docType{package}
\name{manufacturers_googleAuthR}
\alias{manufacturers_googleAuthR}
\alias{manufacturers_googleAuthR-package}
\title{Manufacturer Center API
Public API for managing Manufacturer Center related data.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:01:47
filename: /Users/mark/dev/R/autoGoogleAPI/googlemanufacturersv1.auto/R/manufacturers_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/manufacturercenter
}
}
