% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manufacturers_objects.R
\name{Product}
\alias{Product}
\title{Product Object}
\usage{
Product(manuallyDeletedAttributes = NULL, issues = NULL,
  finalAttributes = NULL, productId = NULL, uploadedAttributes = NULL,
  parent = NULL, manuallyProvidedAttributes = NULL, targetCountry = NULL,
  contentLanguage = NULL, name = NULL)
}
\arguments{
\item{manuallyDeletedAttributes}{Names of the attributes of the product deleted manually via the}

\item{issues}{A server-generated list of issues associated with the product}

\item{finalAttributes}{Final attributes of the product}

\item{productId}{The ID of the product}

\item{uploadedAttributes}{Attributes of the product uploaded via the Manufacturer Center API or via}

\item{parent}{Parent ID in the format `accounts/{account_id}`}

\item{manuallyProvidedAttributes}{Attributes of the product provided manually via the Manufacturer Center UI}

\item{targetCountry}{The target country of the product as a CLDR territory code (for example,}

\item{contentLanguage}{The content language of the product as a two-letter ISO 639-1 language code}

\item{name}{Name in the format `{target_country}:{content_language}:{product_id}`}
}
\value{
Product object
}
\description{
Product Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Product data.
}
